1.53 переписываю на jetpack-compose (itemLoader, main)
