--- app/src/main/java/com/pavlov/MyShadowGallery/ItemLoaderActivity.kt
@@ -1,45 +1,0 @@
-package com.pavlov.MyShadowGallery
-
-import android.Manifest
-import android.app.Dialog
-import android.content.Intent
-import android.content.pm.PackageManager
-import android.net.Uri
-import android.os.Build
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.activity.viewModels
-import androidx.annotation.RequiresApi
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.grid.*
-import androidx.compose.material.*
-import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.asImageBitmap
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
-import androidx.core.app.ActivityCompat
-import androidx.core.content.ContextCompat
-import com.pavlov.MyShadowGallery.ui.MyShadowGalleryTheme
-import com.pavlov.MyShadowGallery.util.showToast
-
-class ItemLoaderActivity : ComponentActivity() {
-
-    private val viewModel: ItemLoaderViewModel by viewModels()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        setContent {
-            MyShadowGalleryTheme {
-                ItemLoaderScreen(viewModel = viewModel, onBack = { finish() })
-            }
-        }
-    }
-}
