+++ app/src/main/java/com/pavlov/MyShadowGallery/security/KeyInputActivity.kt
@@ -16,7 +16,7 @@ import android.widget.ImageButton
 import android.widget.TextView
 import android.widget.Toast
 import androidx.core.widget.doOnTextChanged
-import com.pavlov.MyShadowGallery.MainPageActivity
+import com.pavlov.MyShadowGallery.ui.main.MainPageActivity
 import com.pavlov.MyShadowGallery.R
 import com.pavlov.MyShadowGallery.util.APK
 import com.pavlov.MyShadowGallery.util.APKM
