+++ app/src/main/java/com/pavlov/MyShadowGallery/ui/main/MainPageActivity.kt
@@ -1,9 +1,9 @@
-package com.pavlov.MyShadowGallery
+package com.pavlov.MyShadowGallery.ui.main
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import com.pavlov.MyShadowGallery.ui.MyShadowGalleryTheme
+import com.pavlov.MyShadowGallery.ui.theme.MyShadowGalleryTheme
 
 class MainPageActivity : ComponentActivity() {
