+++ app/src/main/java/com/pavlov/MyShadowGallery/ui/theme/BackgroundImageBase.kt
@@ -1,4 +1,4 @@
-package com.pavlov.MyShadowGallery
+package com.pavlov.MyShadowGallery.ui.theme
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
@@ -9,6 +9,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
+import com.pavlov.MyShadowGallery.R
 
 @Composable
 fun BackgroundImageBase() {
