+++ app/src/main/java/com/pavlov/MyShadowGallery/ui/theme/Typography.kt
@@ -1,4 +1,4 @@
-package com.pavlov.MyShadowGallery
+package com.pavlov.MyShadowGallery.ui.theme
 
 import androidx.compose.material.Typography
